+++ src/framework/components/physics/rigidbody_system.js
@@ -632,8 +632,8 @@ pc.extend(pc, function () {
                     if ((flags0 & pc.BODYFLAG_NORESPONSE_OBJECT) ||
                         (flags1 & pc.BODYFLAG_NORESPONSE_OBJECT)) {
 
-                        var e0Events = e0.collision.hasEvent("triggerenter") || e0.collision.hasEvent("triggerleave");
-                        var e1Events = e1.collision.hasEvent("triggerenter") || e1.collision.hasEvent("triggerleave");
+                        var e0Events = e0.collision ? e0.collision.hasEvent("triggerenter") || e0.collision.hasEvent("triggerleave") : false;
+                        var e1Events = e1.collision ? e1.collision.hasEvent("triggerenter") || e1.collision.hasEvent("triggerleave") : false;
 
                         if (e0Events) {
                             // fire triggerenter events
@@ -654,8 +654,8 @@ pc.extend(pc, function () {
                             }
                         }
                     } else {
-                        var e0Events = e0.collision.hasEvent("collisionstart")  || e0.collision.hasEvent("collisionend")|| e0.collision.hasEvent("contact");
-                        var e1Events = e1.collision.hasEvent("collisionstart") || e1.collision.hasEvent("collisionend") || e1.collision.hasEvent("contact");
+                        var e0Events = e0.collision ? e0.collision.hasEvent("collisionstart")  || e0.collision.hasEvent("collisionend")|| e0.collision.hasEvent("contact") : false;
+                        var e1Events = e1.collision ? e1.collision.hasEvent("collisionstart") || e1.collision.hasEvent("collisionend") || e1.collision.hasEvent("contact") : false;
                         var globalEvents = this.hasEvent("contact");
 
                         if (globalEvents || e0Events || e1Events) {