+++ CHANGES.md
@@ -1,5 +1,19 @@
 # PlayCanvas Engine Changes
 
+### v0.164.3
+* Add pc.Application#timeScale
+
+### v0.164.2
+* [FIX] pc.Mat3 setIdentity was setting 16 values instead of 9
+* [FIX] Errors if rigid body component is removed
+* [FIX] Make sure asset ids are ints
+* [FIX] Documentation fixes
+
+### v0.164.1
+* [FIX] Make sure there is a collision component when checking for collision events.
+* [FIX] Remove camera from system's camera list when the camera component is removed.
+* [FIX] Removed global time variable that prevents multiple applications running on the same page.
+
 ### v0.164.0
 * Added Entity#addComponent and Entity#removeComponent
 * Deprecated ComponentSystem#addComponent and ComponentSystem#removeComponent